+++ v3/examples/hide-window/main.go
@@ -44,7 +44,7 @@ func main() {
 	}
 
 	// Click Dock icon tigger application show
-	app.On(events.Mac.ApplicationShouldHandleReopen, func(event *application.Event) {
+	app.OnApplicationEvent(events.Mac.ApplicationShouldHandleReopen, func(event *application.ApplicationEvent) {
 		println("reopen")
 		window.Show()
 	})